+++ Readme.md
@@ -8,6 +8,9 @@
 ## 서버 프로그램 작업 로그
 
 ### 20170113
+>데이터베이스 매니저 관리를 위해 각 테이블 별로 python 파일을 만들어 분류하기로 정함 우선적으로 Notice(공지), Product(제품), Coupon(쿠폰), CouponShape(쿠폰모양) 에 관련되어 만듬.
+
+### 20170113
 >데이터베이스 연동 중 데이터 추가 부분에 버그를 드디어 고쳤다 
 >commit api를 사용해야 됬었는데 그걸 이제 알음 샹
