+++ webService/BackEnd/DatabaseManager.py
@@ -363,302 +363,6 @@ def InsertCustomerLocationInfo(customerAndStoreRegisteredId, customerLatitude, c
         print "Error in InsertCustomerLocationInfo: " + queryResultData
     return HttpResponse(queryResultData)
 
-#새로운 쿠폰 등록
-def InsertNewCoupon(request):
-    queryResultData = None
-    databaseQuery = None
-
-    try:
-        shopId = request.GET.get('shopId', None)
-        couponId = request.GET.get('couponId', None)
-        couponTitle = request.GET.get('couponTitle', '')
-        couponBody = request.GET.get('couponBody', '')
-        couponShapeIconCode = request.GET.get('couponShopIconCode', '0')
-
-        shopkeeperLatitude = request.GET.get('shopkeeperLatitude', '0.00')
-        shopkeeperLongitude = request.GET.get('shopkeeperLongitude', '0.00')
-        changedDate = request.GET.get('changedDate', '0000-00-00')
-
-        if shopId == None or couponId == None:
-            return HttpResponse("Fail")
-
-        databaseQuery = "insert into `매장쿠폰등록정보` (`매장번호`, `쿠폰고유번호`, `제목`, `내용`, `쿠폰모양코드`) values(" \
-        + shopId + ", '" + couponId + "', '" + couponTitle + "', '" + couponBody + "', " + couponShapeIconCode + ");"
-
-        queryResultData = ExecuteQueryToDatabase(databaseQuery)
-
-        InsertShopkeeperLocationInfo(shopId, shopkeeperLatitude, shopkeeperLongitude, changedDate)
-
-    except:
-        print "Error in InsertNewCoupon: " + queryResultData
-
-    return HttpResponse(queryResultData)
-
-#업로드한 쿠폰 정보를 변경
-def UpdateUploadedCoupon(request):
-    queryResultData = None
-    databaseQuery = None
-
-    try:
-        shopId = request.GET.get('shopId', None)
-        couponId = request.GET.get('couponId', None)
-        couponTitle = request.GET.get('couponTitle', '')
-        couponBody = request.GET.get('couponBody', '')
-        couponShapeIconCode = request.GET.get('couponShopIconCode', '0')
-
-        shopkeeperLatitude = request.GET.get('shopkeeperLatitude', '0.00')
-        shopkeeperLongitude = request.GET.get('shopkeeperLongitude', '0.00')
-        changedDate = request.GET.get('changedDate', '0000-00-00')
-
-        if shopId == None or couponId == None:
-            return HttpResponse("Fail")
-
-        databaseQuery = "update `매장쿠폰등록정보`" \
-        + " set `제목` = '" + couponTitle + "', `내용` = '" + couponBody + "', `쿠폰모양코드` = " + couponShapeIconCode \
-        + " where `매장번호` = " + shopId + " and `쿠폰고유번호` = '" + couponId + "';"
-
-        queryResultData = ExecuteQueryToDatabase(databaseQuery)
-
-        InsertShopkeeperLocationInfo(shopId, shopkeeperLatitude, shopkeeperLongitude, changedDate)
-
-    except:
-        print "Error in UpdateUploadedCoupon: " + queryResultData
-
-    return HttpResponse(queryResultData)
-
-#업로드한 쿠폰을 삭제
-def DelUploadedCoupon(request):
-    queryResultData = None
-    databaseQuery = None
-
-    try:
-        shopId = request.GET.get('shopId', None)
-        couponId = request.GET.get('couponId', None)
-
-        shopkeeperLatitude = request.GET.get('shopkeeperLatitude', '0.00')
-        shopkeeperLongitude = request.GET.get('shopkeeperLongitude', '0.00')
-        changedDate = request.GET.get('changedDate', '0000-00-00')
-
-        if shopId == None or couponId == None:
-            return HttpResponse("Fail")
-
-        databaseQuery = "update `매장쿠폰등록정보`" \
-        + " set `삭제 여부` = 1" \
-        + " where `매장번호` = " + shopId + " and `쿠폰고유번호` = '" + couponId + "';"
-
-        queryResultData = ExecuteQueryToDatabase(databaseQuery)
-
-        InsertShopkeeperLocationInfo(shopId, shopkeeperLatitude, shopkeeperLongitude, changedDate)
-
-    except:
-        print "Error in DelUploadedCoupon: " + queryResultData
-
-    return HttpResponse(queryResultData)
-
-def InsertShopkeeperLocationInfo(storeId, shopkeeperLatitude, shopkeeperLongitude, changedDate):
-    queryResultData = None
-    databaseQuery = None
-    try :
-        databaseQuery = "insert into `업로더 위치 정보` values(" + storeId + ", " + shopkeeperLatitude \
-        + ", " + shopkeeperLongitude + ", " + changedDate + ");"
-
-        queryResultData = ExecuteQueryToDatabase(databaseQuery)
-    except:
-        print "Error in InsertCustomerLocationInfo: " + queryResultData
-    return HttpResponse(queryResultData)
-
-#제품 추가
-def InsertNewProductName(request):
-    queryResultData = None
-    databaseQuery = None
-    try:
-        shopId = request.GET.get('shopId', None)
-        productId = request.GET.get('productId', None)
-        productName = request.GET.get('productName', None)
-        if productName == None or shopId == None:
-            return HttpResponse("Fail")
-
-        databaseQuery = "insert into `매장 제품 정보` (`매장번호`, `제품코드`, `이름`) " \
-        + "select * from (select " + shopId + ", " + productId + ", '" + productName + "') as compareTemp " \
-        + "where not exists (" \
-        + "select `매장번호`, `제품코드` from `매장 제품 정보` where `매장번호` = " + shopId + " and " \
-        + "`제품코드` = " + productId + ") limit 1;"
-
-        queryResultData = ExecuteQueryToDatabase(databaseQuery)
-
-        #return HttpResponse("ok")
-    except:
-        print "Error in InsertNewProductName: " + queryResultData
-    return HttpResponse(queryResultData)
-
-#제품 수정
-def UpdateRegisteredProductName(request):
-    queryResultData = None
-    databaseQuery = None
-
-    try:
-        shopId = request.GET.get('shopId', None)
-        productId = request.GET.get('productId', None)
-        newProductName = request.GET.get('newProductName', None)
-
-        if shopId == None or productId == None or newProductName == None:
-            return HttpResponse("Fail")
-
-        databaseQuery = "update `매장 제품 정보` set `이름` = '" + newProductName + "'" \
-        + " where `매장번호` = " + shopId + " and `productId` = " + productId + ";"
-
-        queryResultData = ExecuteQueryToDatabase(queryResultData)
-        #return HttpResponse("ok")
-    except:
-        print "Error in UpdateRegisteredProductName: " + queryResultData
-    return HttpResponse(queryResultData)
-
-#제품 삭제
-def DelRegisteredProductName(request):
-    queryResultData = None
-    databaseQuery = None
-
-    try:
-        shopId = request.GET.get('shopId', None)
-        productId = request.GET.get('productId', None)
-
-        if shopId == None or productId == None :
-            return HttpResponse("Fail")
-
-        databaseQuery = "update `매장 제품 정보` set `삭제 여부` = 1" \
-        + " where `매장번호` = " + shopId + " and `productId` = " + productId + ";"
-
-        queryResultData = ExecuteQueryToDatabase(queryResultData)
-        #return HttpResponse("ok")
-    except:
-        print "Error in UpdateRegisteredProductName: " + queryResultData
-    return HttpResponse(queryResultData)
-
-#새로운 공지사항 추가
-def InsertNewStoreNoticeInfo(request):
-    queryResultData = None
-    databaseQuery = None
-
-    try:
-        shopId = request.GET.get('shopId', None)
-        noticeTitle = request.GET.get('noticeTitle', '')
-        noticeBody = request.GET.get('noticeBody', '')
-        noticeStartDate = request.GET.get('noticeStartDate', '0000-00-00')
-        noticeStopDate = request.GET.get('noticeStopDate', '0000-00-00')
-        noticeLastUpdateDate = request.GET.get('noticeLastUpdateDate', '0000-00-00')
-
-        if shopId == None:
-            return HttpResponse("Fail")
-
-        databaseQuery = "insert into `매장공지 정보` (`매장번호`, `제목`, `내용`, `공지 시작 날짜`, `공지 마감 날짜`, `마지막 편집 날짜`)" \
-        + " values (" + shopId + ", '" + noticeTitle + "', '" + noticeBody + "', '" + noticeStartDate + "', '" + noticeStopDate + "', '" \
-        + noticeLastUpdateDate + "');"
-
-        queryResultData = ExecuteQueryToDatabase(databaseQuery)
-    except:
-        print "Error in InsertNewStoreNoticeInfo: " + queryResultData
-    return HttpResponse(queryResultData)
-
-#기존의 공지사항 편집
-def UpdateStoreNoticeInfo(request):
-    queryResultData = None
-    databaseQuery = None
-
-    try:
-        shopId = request.GET.get('shopId', None)
-        noticeId = request.GET.get('noticeId', None)
-        noticeTitle = request.GET.get('noticeTitle', '')
-        noticeBody = request.GET.get('noticeBody', '')
-        noticeStartDate = request.GET.get('noticeStartDate', '0000-00-00')
-        noticeStopDate = request.GET.get('noticeStopDate', '0000-00-00')
-        noticeLastUpdateDate = request.GET.get('noticeLastUpdateDate', '0000-00-00')
-
-        if shopId == None and noticeId == None:
-            return HttpResponse("Fail")
-
-        databaseQuery = "update `매장공지 정보`" \
-        + " set `제목` = '" + noticeTitle + "', `내용` = '" + noticeBody + "', `공지 시작 날짜` = '" + noticeStartDate + "', " \
-        + "`공지 마감 날짜` = '" + noticeStopDate + "', `마지막 편집 날짜` = '" + noticeLastUpdateDate + "' " \
-        + "where `매장번호` = " + shopId + " and `공지번호` = " + noticeId + ";"
-
-        queryResultData = ExecuteQueryToDatabase(databaseQuery)
-
-    except:
-        print "Error in UpdateStoreNoticeInfo: " + queryResultData
-    return HttpResponse(queryResultData)
-
-#기존의 공지사항 삭제
-def DelStoreNoticeInfo(request):
-    queryResultData = None
-    databaseQuery = None
-
-    shopId = request.GET.get('shopId', None)
-    noticeId = request.GET.get('noticeId', None)
-
-    try :
-        databaseQuery = "update `매장공지 정보`"\
-                        + "set `삭제 여부` = 1" \
-                        + "where `매장번호` = " + shopId + " and `공지번호` = " + noticeId + ";"
-        queryResultData = ExecuteQueryToDatabase(databaseQuery)
-    except :
-        print "Error in DelStoreNoticeInfo: " + queryResultData
-
-    return HttpResponse(queryResultData)
-
-def InsertCouponShapeInfo(request) :
-    couponShapeCode = request.GET.get('code', None)
-    couponImageAddress = request.GET.get('address', None)
-    couponShapePrice = request.GET.get('price', '0')
-    couponShapeLimitTime = request.GET.get('limit', '-1')
-    couponShapeEx = request.GET.get('ex', None)
-
-    databaseQuery = None
-    queryResultData = None
-
-    try:
-        databaseQuery = "insert into `쿠폰모양 정보` values(" + str(couponShapeCode) + ", " + str(couponImageAddress) \
-                        + ", "+ str(couponShapePrice) + ", " + str(couponShapeLimitTime) + ", " + str(couponShapeEx) + ");"
-
-        queryResultData = ExecuteQueryToDatabase(databaseQuery)
-    except:
-        print "Error in InsertCouponShapeInfo: " + queryResultData
-    return HttpResponse(queryResultData)
-
-def UpdateCouponShapeInfo(request) :
-    couponShapeCode = request.GET.get('code', None)
-    couponImageAddress = request.GET.get('address', None)
-    couponShapePrice = request.GET.get('price', '0')
-    couponShapeLimitTime = request.GET.get('limit', '-1')
-    couponShapeEx = request.GET.get('ex', None)
-
-    databaseQuery = None
-    queryResultData = None
-
-    try:
-        databaseQuery = "update `쿠폰모양 정보` set `쿠폰 이미지 저장 경로` = " + str(couponImageAddress) + ", `쿠폰 모양 가격` = " + str(couponShapePrice) \
-                        + ", `쿠폰 모양 기간` = " + str(couponShapeLimitTime) + ", `쿠폰 모양 설명` = " + str(couponShapeEx) \
-                        + "where `쿠폰모양코드` == " + str(couponShapeCode) + ";"
-        queryResultData = ExecuteQueryToDatabase(databaseQuery)
-    except:
-        print "Error in UpdateCouponShapeInfo: " + queryResultData
-    return HttpResponse(queryResultData)
-
-def InsertCouponShapeCollectLog(request) :
-    myStoreId = request.GET.get('storeId','0')
-    couponShapeCode = request.GET.get('code', None)
-    couponEditTime = request.GET.get('editTime', None)
-
-    databaseQuery = None
-    queryResultData = None
-
-    try:
-        databaseQuery = "insert into `매장 쿠폰 모양 수집 로그` values(" + str(myStoreId) + ", " + str(couponShapeCode) + ", " + str(couponEditTime) + ");"
-        queryResultData = ExecuteQueryToDatabase(databaseQuery)
-    except:
-        print "Error in InsertCouponShapeCollectLog: " + queryResultData
-    return HttpResponse(queryResultData)
-
-
 def InsertProductOptimalStock(request) :
     productCode = request.GET.get('productCode', None)
     engineVersion = request.GET.get('engineVersion', None) #엔진버전
